refactore interleaved utils into its own module
