+++ src/parser.rs
@@ -321,49 +321,6 @@ pub enum Indices {
     NumIndices(u32),
 }
 
-fn copy_to_interleaved_array<T: Copy + Send + Sync>(dst: &mut [T], src: &[T], byte_offset: usize, byte_stride: usize, begin: usize, end: usize) {
-    debug_assert_eq!(byte_offset % size_of::<T>(), 0);
-    debug_assert_eq!(byte_stride % size_of::<T>(), 0);
-
-    let offset = byte_offset / size_of::<T>();
-    let stride = byte_stride / size_of::<T>();
-
-    for (dst, src) in dst[offset..].iter_mut().step_by(stride).zip(&src[begin..end]) {
-        *dst = *src
-    }
-}
-
-
-fn fill_to_interleaved_array<T: Copy + Send + Sync>(dst: &mut [T], src: T, byte_offset: usize, byte_stride: usize, begin: usize, end: usize) {
-    debug_assert_eq!(byte_offset % size_of::<T>(), 0);
-    debug_assert_eq!(byte_stride % size_of::<T>(), 0);
-
-    let offset = byte_offset / size_of::<T>();
-    let stride = byte_stride / size_of::<T>();
-
-    let end = (offset + stride * (end - begin)).min(dst.len());
-
-    for dst in dst[offset..end].iter_mut().step_by(stride) {
-        *dst = src;
-    }
-}
-
-
-#[test]
-fn test_copy_interleaved() {
-    let src = [0f32, 1., 2., 3., 4., 5., 6.];
-    let mut dst = [0.; 14];
-    copy_to_interleaved_array(&mut dst, &src, size_of::<f32>() * 2, size_of::<f32>() * 2, 1, 7);
-    assert_eq!(dst, [0., 0., 1., 0., 2., 0., 3., 0., 4., 0., 5., 0., 6., 0.]);
-}
-
-#[test]
-fn test_fill_interleaved() {
-    let src = 1.;
-    let mut dst = [0f32; 14];
-    fill_to_interleaved_array(&mut dst, src, size_of::<f32>() * 2, size_of::<f32>() * 2, 1, 7);
-    assert_eq!(dst, [0., 0., 1., 0., 1., 0., 1., 0., 1., 0., 1., 0., 1., 0.]);
-}
 
 // use crate::types_2_0::*;
 // use crate::reader_2_0::*;
@@ -892,7 +849,7 @@ macro_rules! impl_parser {
                                     let offset = attrib_offsets[attrib] as usize + c * bytes_per_element;
                                     match attrib {
                                         Attribute::MaterialIndex =>
-                                            super::fill_to_interleaved_array(
+                                            fill_to_interleaved_array(
                                                 bytemuck::cast_slice_mut(dst),
                                                 sub_mesh.material_index,
                                                 offset,
@@ -901,7 +858,7 @@ macro_rules! impl_parser {
                                                 sub_mesh.vertices.len as usize,
                                             ),
                                         Attribute::ObjectId =>
-                                            super::fill_to_interleaved_array(
+                                            fill_to_interleaved_array(
                                                 bytemuck::cast_slice_mut(dst),
                                                 sub_mesh.object_id,
                                                 offset,
@@ -952,7 +909,7 @@ macro_rules! impl_parser {
                                     .fill(sub_mesh.object_id);
                             }
 
-                            super::copy_to_interleaved_array::<i16>(
+                            copy_to_interleaved_array::<i16>(
                                 bytemuck::cast_slice_mut(&mut position_buffer),
                                 unsafe{ sub_mesh.vertices.position.x.as_slice(sub_mesh.vertices.len) },
                                 vertex_offset * position_stride + 0,
@@ -961,7 +918,7 @@ macro_rules! impl_parser {
                                 sub_mesh.vertices.len as usize,
                             );
 
-                            super::copy_to_interleaved_array::<i16>(
+                            copy_to_interleaved_array::<i16>(
                                 bytemuck::cast_slice_mut(&mut position_buffer),
                                 unsafe{ sub_mesh.vertices.position.y.as_slice(sub_mesh.vertices.len) },
                                 vertex_offset * position_stride + 2,
@@ -970,7 +927,7 @@ macro_rules! impl_parser {
                                 sub_mesh.vertices.len as usize,
                             );
 
-                            super::copy_to_interleaved_array::<i16>(
+                            copy_to_interleaved_array::<i16>(
                                 bytemuck::cast_slice_mut(&mut position_buffer),
                                 unsafe{ sub_mesh.vertices.position.z.as_slice(sub_mesh.vertices.len) },
                                 vertex_offset * position_stride + 4,
